# TP1_API_A25/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Films API",
        "version": "1.0"
    },
    "paths": {
        "/api/critics/{id}": {
            "delete": {
                "tags": [
                    "Critics"
                ],
                "summary": "Deletes a critic",
                "operationId": "ec92f7c003d7a3e6dda0590f39d77589",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "critic id",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/films": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Gets list of films",
                "operationId": "35226c835d51209245133d69fff773d7",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/films/{id}/actors": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Gets actors of a film",
                "operationId": "cdf8c200bb7d2869ca32bb44097049d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "film id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/films/{id}/critiques": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Gets a film with its critics",
                "operationId": "6c55d863092c7dd99806fd5e41599e0c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "film id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/films/{id}/moyenne-critiques": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Gets average score of a film",
                "operationId": "e293a6127b1ea72353ab5055b13f7930",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "film id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/films/recherche": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Search films with criteria",
                "operationId": "45d35f694ede2daba12be9648003561e",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Filter by title keyword"
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "description": "Filter by rating"
                    },
                    {
                        "name": "minLength",
                        "in": "query",
                        "description": "Minimum length"
                    },
                    {
                        "name": "maxLength",
                        "in": "query",
                        "description": "Maximum length"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Creates a user",
                "operationId": "ff5bd3e20fb2fec4d8bcd3b9f1a0352a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "login": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "language_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Updates a user",
                "operationId": "39a41be7e31a396191eb7705bfd7831b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "login": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "language_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/users/{id}/preferred-language": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Gets preferred language of a user",
                "operationId": "296af2f75065672eb87973c9d60c8cdc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Critics",
            "description": "Critics"
        },
        {
            "name": "Films",
            "description": "Films"
        },
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}